+++ include/cuda/kernels/vector_reduction/vector_reduction_recursive.cuh
@@ -76,7 +76,7 @@ __global__ void vector_reduction_by_blocks_parallel(
     }
 
     // Warp-shuffle reduction
-    for (int reduced_lanes = 1; reduced_lanes < WARP_SIZE; reduced_lanes <<= 1) {
+    for (unsigned short reduced_lanes = 1; reduced_lanes < WARP_SIZE; reduced_lanes <<= 1) {
         value = Operation::apply(value, __shfl_down_sync(__activemask(), value, reduced_lanes));
     }
 
@@ -93,7 +93,7 @@ __global__ void vector_reduction_by_blocks_parallel(
         value = (tid_warp < shm_size) ? shm[tid_warp] : Number(0);
 
         // Warp-shuffle reduction: compute total value for this block
-        for (int reduced_lanes = 1; reduced_lanes < WARP_SIZE; reduced_lanes <<= 1) {
+        for (unsigned short reduced_lanes = 1; reduced_lanes < WARP_SIZE; reduced_lanes <<= 1) {
             value = Operation::apply(value, __shfl_down_sync(__activemask(), value, reduced_lanes));
         }
